[Doc V3] Update PDF search function
